requirements.txt
dj-database-url==0.4.2
Django==1.11.5
django-filter==1.0.4
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
gunicorn==19.7.1
psycopg2==2.7.3.1
pytz==2017.2
whitenoise==3.3.0
